+++ frontend/app/api/auth/signup/route.ts
@@ -1,5 +1,5 @@
 import { signUpSchema } from '@/lib/schema';
-import { apiRequest } from '@/lib/utils';
+import { apiRequest } from '@/lib/server/apiRequest';
 import { ZodError } from 'zod';
 
 // /api/auth/signup
